requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
dataforseo-client
starlette>=0.38.0
pydantic>=2.7.0

# AI & ML Libraries
openai>=1.30.0
anthropic>=0.40.0

# Database
supabase>=2.2.0
psycopg2-binary

# Data Science & Clustering  
numpy>=1.26.0
scikit-learn>=1.4.0
pandas>=2.0.0
hdbscan>=0.8.33

# HTTP & Async
aiohttp>=3.9.0
requests>=2.31.0
httpx>=0.25.0

# Templates & Static
jinja2>=3.1.2